cloud_functions/waitlist/requirements.txt
functions-framework==3.5.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-firestore==2.13.1

